+++ implementation/js/src/infrastructure/lib+-/marionette/item-view.js
@@ -1,15 +1,105 @@
 /**
  * Marionette.ItemView Enhancements (can be used in Layout as well) - Note that you can NOT use these in a CompositeView.
  *
- * 1. SVG (view:fit-paper, view:paper-resized, view:paper-ready)
- * 2. Basic Editors (view as form piece)
- * 3. Render with data (view:render-data, view:data-rendered)
+ * 0. actions
+ * 1. svg (view:fit-paper, view:paper-resized, view:paper-ready)
+ * 2. basic Editors (view as form piece)
+ * 3. tooltips
+ * 4. overlay
+ * 5. data event listener (view:render-data, view:data-rendered)
  *
  * @author Tim.Liu
- * @create 2014.02.26
+ * @created 2014.02.26
+ * @updated 2015.08.03
  */
 
 ;(function(app){
+
+	/**
+	 * Action Tag listener hookups +actions{} (do it in initialize())
+	 * + event forwarding ability to action tags
+	 * Usage:
+	 * 		1. add action tags to html template -> e.g <div ... action="method name or *:event name"></div> 
+	 * 		2. implement the action method name in UI definition body's actions{} object. 
+	 * 		functions under actions{} are invoked with 'this' as scope (the view object).
+	 * 		functions under actions{} are called with a 2 params ($action, e) which is a jQuery object referencing the action tag and the jQuery prepared event object, use e.originalEvent to get the DOM one.
+	 *
+	 * Options
+	 * -------
+	 * 1. uiName - [UNKNOWN.View] this is optional, mainly for better debugging msg;
+	 * 2. passOn - [false] this is to let the clicking event of action tags bubble up if an action listener is not found. 
+	 *
+	 * Note:
+	 * A. We removed _.bind() altogether from the enableActionTags() function and use Function.apply(scope, args) instead for listener invocation to avoid actions{} methods binding problem.
+	 * Functions under actions will only be bound ONCE to the first instance of the view definition, since _.bind() can not rebind functions that were already bound, other instances of
+	 * the view prototype will have all the action listeners bound to the wrong view object. This holds true to all nested functions, if you assign the bound version of the function back to itself
+	 * e.g. this.nest.func = _.bind(this.nest.func, this); - Do NOT do this in initialize()/constructor()!! Use Function.apply() for invocation instead!!!
+	 *
+	 * B. We only do e.stopPropagation for you, if you need e.preventDefault(), do it yourself in the action impl;
+	 */
+	_.extend(Backbone.Marionette.ItemView.prototype, {
+
+		enableActionTags: function(uiName, passOn){ //the uiName is just there to output meaningful dev msg if some actions haven't been implemented.
+
+			if(_.isBoolean(uiName)){
+				passOn = uiName;
+				uiName = '';
+			}
+			passOn = passOn || false;
+			this.events = this.events || {};
+			//add general action tag clicking event and listener
+			_.extend(this.events, {
+				'click [action]': '_doAction'
+			});
+			this.actions = this.actions || {}; 	
+			uiName = uiName || this.name || 'UNKNOWN.View';
+
+			this._doAction = function(e){
+
+				var $el = $(e.currentTarget);
+				var action = $el.attr('action') || 'UNKNOWN';
+				var lockTopic = $el.attr('lock'),
+				unlockTopic = $el.attr('unlock');
+
+				if(unlockTopic) app.unlock(unlockTopic);
+
+				if(lockTopic && !app.lock(lockTopic)){
+					e.stopPropagation();
+					e.preventDefault();
+					app.trigger('app:blocked', action, lockTopic);
+					return;
+				}
+
+				if($el.hasClass('disabled') || $el.parent().hasClass('disabled')) {
+					e.stopPropagation();
+					e.preventDefault();					
+					return;
+				}
+
+				//allow triggering certain event only.
+				var eventForwarding = String(action).split(':');
+				if(eventForwarding.length >= 2) {
+					eventForwarding.shift();
+					e.stopPropagation(); //Important::This is to prevent confusing the parent view's action tag listeners.
+					return this.trigger(eventForwarding.join(':'));
+				}
+
+				var doer = this.actions[action];
+				if(doer) {
+					e.stopPropagation(); //Important::This is to prevent confusing the parent view's action tag listeners.
+					doer.apply(this, [$el, e, lockTopic]); //use 'this' view object as scope when applying the action listeners.
+				}else {
+					if(passOn){
+						return;
+					}else {
+						e.stopPropagation(); //Important::This is to prevent confusing the parent view's action tag listeners.
+					}
+					throw new Error('DEV::' + (uiName || 'UI Component') + '::You have not yet implemented this action - [' + action + ']');
+				}
+			};		
+		},
+			
+	});
 
 	/**
 	 * Inject a svg canvas within view. - note that 'this' in cb means paper.
@@ -216,6 +306,61 @@
 
 			//auto setValues according to this.model?
 			
+		}
+
+	});
+
+	/**
+	 * Enable Tooltips (do it in initialize())
+	 * This is used for automatically activate tooltips after render
+	 *
+	 * Options
+	 * -------
+	 * bootstrap tooltip config
+	 */
+
+	_.extend(Backbone.Marionette.ItemView.prototype, {
+
+		enableTooltips: function(options){
+			this.listenTo(this, 'render', function(){
+				//will activate tooltip with specific options object - see /libs/bower_components/bootstrap[x]/docs/javascript.html#tooltips
+				this.$('[data-toggle="tooltip"]').tooltip(options);
+			});
+		}
+
+	});
+
+	/**
+	 * Overlay
+	 * options:
+	 * 1. anchor - css selector of parent html el
+	 * 2. rest of the $.overlay plugin options without content and onClose
+	 */
+	_.extend(Backbone.Marionette.ItemView.prototype, {
+
+		enableOverlay: function(){
+			this._overlayConfig = _.isBoolean(this.overlay)? {}: this.overlay;
+			this.overlay = function(options){
+				options = options || {};
+				var $anchor = $(options.anchor || 'body');
+				var that = this;
+				this.listenTo(this, 'close', function(){
+					$anchor.overlay();//close the overlay if this.close() is called.
+				});
+				$anchor.overlay(_.extend(this._overlayConfig, options, {
+					content: function(){
+						return that.render().el;
+					},
+					onShow: function(){
+						//that.trigger('show'); //Trigger 'show' doesn't invoke onShow, use triggerMethod the Marionette way!
+						that.triggerMethod('show'); //trigger event while invoking on{Event};
+					},
+					onClose: function(){
+						that.close(); //closed by overlay x
+					}
+				}));
+				return this;
+			};			
 		}
 
 	});