+++ implementation/js/src/infrastructure/lib+-/marionette/view.js
@@ -2,136 +2,29 @@
  * Here we extend the html tag attributes to be auto-recognized by a Marionette.View.
  * This simplifies the view creation by indicating added functionality through template string. (like angular.js?)
  *
- * Optional
- * --------
- * 1. action tags auto listener hookup with mutex-locking
- * 2. tooltip
- * 3. overlay - use this view as an overlay
- * 
- *
  * Fixed
  * -----
- * auto ui tags detect and register.
- * +meta event programming
- * 	view:* (event-name) - on* (camelized)
+ * 0. shiv empty template.
+ * 1. auto ui tags detection in template.
+ * 2. +meta event programming
+ * 	view:* (event-name) <--> on* (camelized)
+ * 3. global coop events.
  *
  * 
  * @author Tim.Liu
- * @create 2014.02.25 
+ * @created 2014.02.25
+ * @updated 2015.08.03
  */
 
 
 ;(function(app){
 
-
-/**
- * Action Tag listener hookups +actions{} (do it in initialize())
- * + event forwarding ability to action tags
- * Usage:
- * 		1. add action tags to html template -> e.g <div ... action="method name or *:event name"></div> 
- * 		2. implement the action method name in UI definition body's actions{} object. 
- * 		functions under actions{} are invoked with 'this' as scope (the view object).
- * 		functions under actions{} are called with a 2 params ($action, e) which is a jQuery object referencing the action tag and the jQuery prepared event object, use e.originalEvent to get the DOM one.
- *
- * Options
- * -------
- * 1. uiName - [UNKNOWN.View] this is optional, mainly for better debugging msg;
- * 2. passOn - [false] this is to let the clicking event of action tags bubble up if an action listener is not found. 
- *
- * Note:
- * A. We removed _.bind() altogether from the enableActionTags() function and use Function.apply(scope, args) instead for listener invocation to avoid actions{} methods binding problem.
- * Functions under actions will only be bound ONCE to the first instance of the view definition, since _.bind() can not rebind functions that were already bound, other instances of
- * the view prototype will have all the action listeners bound to the wrong view object. This holds true to all nested functions, if you assign the bound version of the function back to itself
- * e.g. this.nest.func = _.bind(this.nest.func, this); - Do NOT do this in initialize()/constructor()!! Use Function.apply() for invocation instead!!!
- *
- * B. We only do e.stopPropagation for you, if you need e.preventDefault(), do it yourself in the action impl;
- */
 	_.extend(Backbone.Marionette.View.prototype, {
-
-		enableActionTags: function(uiName, passOn){ //the uiName is just there to output meaningful dev msg if some actions haven't been implemented.
-
-			if(_.isBoolean(uiName)){
-				passOn = uiName;
-				uiName = '';
-			}
-			passOn = passOn || false;
-			this.events = this.events || {};
-			//add general action tag clicking event and listener
-			_.extend(this.events, {
-				'click [action]': '_doAction'
-			});
-			this.actions = this.actions || {}; 	
-			uiName = uiName || this.name || 'UNKNOWN.View';
-
-			this._doAction = function(e){
-
-				var $el = $(e.currentTarget);
-				var action = $el.attr('action') || 'UNKNOWN';
-				var lockTopic = $el.attr('lock'),
-				unlockTopic = $el.attr('unlock');
-
-				if(unlockTopic) app.unlock(unlockTopic);
-
-				if(lockTopic && !app.lock(lockTopic)){
-					e.stopPropagation();
-					e.preventDefault();
-					app.trigger('app:blocked', action, lockTopic);
-					return;
-				}
-
-				if($el.hasClass('disabled') || $el.parent().hasClass('disabled')) {
-					e.stopPropagation();
-					e.preventDefault();					
-					return;
-				}
-
-				//allow triggering certain event only.
-				var eventForwarding = String(action).split(':');
-				if(eventForwarding.length >= 2) {
-					eventForwarding.shift();
-					e.stopPropagation(); //Important::This is to prevent confusing the parent view's action tag listeners.
-					return this.trigger(eventForwarding.join(':'));
-				}
-
-				var doer = this.actions[action];
-				if(doer) {
-					e.stopPropagation(); //Important::This is to prevent confusing the parent view's action tag listeners.
-					doer.apply(this, [$el, e, lockTopic]); //use 'this' view object as scope when applying the action listeners.
-				}else {
-					if(passOn){
-						return;
-					}else {
-						e.stopPropagation(); //Important::This is to prevent confusing the parent view's action tag listeners.
-					}
-					throw new Error('DEV::' + (uiName || 'UI Component') + '::You have not yet implemented this action - [' + action + ']');
-				}
-			};		
-		},
-
-			
-	});
-
-
-	/**
-	 * Enable Tooltips (do it in initialize())
-	 * This is used for automatically activate tooltips after render
-	 *
-	 * Options
-	 * -------
-	 * bootstrap tooltip config
-	 */
-
-	_.extend(Backbone.Marionette.View.prototype, {
-
-		enableTooltips: function(options){
-			this.listenTo(this, 'render', function(){
-				//will activate tooltip with specific options object - see /libs/bower_components/bootstrap[x]/docs/javascript.html#tooltips
-				this.$('[data-toggle="tooltip"]').tooltip(options);
-			});
+		isInDOM: function(){
+			if(!this.$el) return false;
+			return $.contains(document.documentElement, this.$el[0]);
 		}
-
 	});
-
 
 	/**
 	 * Fixed enhancement
@@ -139,12 +32,14 @@
 	 * +meta event programming
 	 * 	view:* (event-name) - on* (camelized)
 	 *
-	 * Override View.constructor to affect only decendents, e.g ItemView and CollectionView... (This is the Backbone way of extend...)
+	 * Override View.constructor to affect only decendents, e.g ItemView and CollectionView... 
+	 * (This is the Backbone way of extend...)
 	 * 
 	 */
 	Backbone.Marionette.View.prototype.constructor = function(options){
 		options = options || {};
 
+		//----------------------fixed enhancements--------------------------
 		//fix default tpl to be ' '.
 		this.template = options.template || this.template || ' ';
 
@@ -163,46 +58,54 @@
 
 		//meta-event programming ability
 		app.Util.addMetaEvent(this, 'view');
-		//auto detect and enable view enhancements: actions, [paper(SVG), editors - in item-view enhancement]
-		if(this.actions) this.enableActionTags(this.actions._bubble);
+
+		//global co-op (global events forwarding through app)
+		if(this.coop) {
+			this._postman = {};
+			//register
+			_.each(this.coop, function(e){
+				var self = this;
+				this._postman[e] = function(options){
+					self.trigger('view:' + e, options);
+					//considering the parent-DOM-removed edge case
+					if(!self.isInDOM())
+						app.off('app:coop:' + e, self._postman[e]);
+				};
+				app.on('app:coop:' + e, this._postman[e]);
+			}, this);
+			//cleanup
+			this.listenTo(this, 'close', function(){
+				_.each(this._postman, function(fn, e){
+					app.off('app:coop:' + e, fn);
+				});
+			});
+		}		
+		
+		//---------------------optional view enhancements-------------------
+		//actions (1-click uis)
+		if(this.actions && this.enableActionTags) 
+			this.enableActionTags(this.actions._bubble);
+		
+		//editors
 		if(this.editors && this.activateEditors) this.listenTo(this, 'render', function(){
 			this.activateEditors(this.editors);
 		});
+
+		//svg (if rapheal.js is present)
 		if(this.svg && this.enableSVG) {
 			this.listenTo(this, 'render', this.enableSVG);
 		}
-		if(this.tooltips) {
+
+		//tooltip
+		if(this.tooltips && this.enableTooltips) {
 			this.enableTooltips(this.tooltips);
 		}
-		if(this.overlay){ //give this view the overlaying ability
-			this._overlayConfig = _.isBoolean(this.overlay)? {}: this.overlay;
-			this.overlay = function(options){
-				/**
-				 * options:
-				 * 1. anchor - css selector of parent html el
-				 * 2. rest of the $.overlay plugin options without content and onClose
-				 */
-				options = options || {};
-				var $anchor = $(options.anchor || 'body');
-				var that = this;
-				this.listenTo(this, 'close', function(){
-					$anchor.overlay();//close the overlay if this.close() is called.
-				});
-				$anchor.overlay(_.extend(this._overlayConfig, options, {
-					content: function(){
-						return that.render().el;
-					},
-					onShow: function(){
-						//that.trigger('show'); //Trigger 'show' doesn't invoke onShow, use triggerMethod the Marionette way!
-						that.triggerMethod('show'); //trigger event while invoking on{Event};
-					},
-					onClose: function(){
-						that.close(); //closed by overlay x
-					}
-				}));
-				return this;
-			};
+
+		//overlay (use this view as overlay)
+		if(this.overlay && this.enableOverlay){
+			this.enableOverlay();
 		}
+
 		//auto-enable i18n
 		if(I18N.locale) {
 			this.listenTo(this, 'render', function(){