+++ CHANGELOG.md
@@ -1,14 +1,16 @@
 Change Log
 ==========
-1.8.0 (2015-07-29*)
+1.8.0 (2015-08-03*)
 -------------------
 1. Refined Reusable module method register().
 2. Added _.isPlainObject() to underscore.
+3. Back-ported global co-op e.
+4. Added view.isInDOM() check.
+5. Back-ported dynamic view injection (synced).
+6. Added app.has/get apis.
 
 Working on...
 
-- Back-porting global co-op e
-- Back-porting dynamic view injection 
 - Adding /tests